+++ src/Shared/SqlProcessor.cs
@@ -28,6 +28,7 @@ internal static class SqlProcessor
     private const char NewLineChar = '\n';
     private const char CarriageReturnChar = '\r';
     private const char TabChar = '\t';
+    private const char UnicodePrefixChar = 'N';
 
     private static readonly ConcurrentDictionary<string, SqlStatementInfo> Cache = new();
 
@@ -672,6 +673,10 @@ internal static class SqlProcessor
                 return true;
             }
 
+            // Is the string literal of the form `N'foo'` (i.e. a Unicode literal)?
+            // If so, we want to skip the Unicode prefix when sanitizing.
+            bool isUnicode = state.ParsePosition >= 1 && sql[state.ParsePosition - 1] is UnicodePrefixChar;
+
             // Use index arithmetic instead of slicing
             var searchPos = state.ParsePosition + 1;
             while (searchPos < sql.Length)
@@ -686,6 +691,12 @@ internal static class SqlProcessor
                     }
 
                     // Found terminating quote
+                    if (isUnicode)
+                    {
+                        // Skip the Unicode prefix by overwriting the previous position instead
+                        state.SanitizedPosition--;
+                    }
+
                     state.ParsePosition = searchPos + 1;
                     buffer[state.SanitizedPosition++] = SanitizationPlaceholder;
                     return true;